requirements.txt
aiobotocore==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
botocore==1.42.55
bounded-pool-executor==0.0.3
branca==0.8.2
cachetools==6.2.6
certifi==2026.2.25
cftime==1.6.5
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
earthaccess==0.16.0
folium==0.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
geopandas==1.1.2
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.1
multimethod==2.0.2
narwhals==2.17.0
netCDF4==1.7.4
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.5.2
pqdm==0.2.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-cmr==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3fs==2026.2.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.54.0
tenacity==9.1.4
tinynetrc==1.3.1
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wrapt==2.1.1
xarray==2026.2.0
xyzservices==2025.11.0
yarl==1.22.0
